A few changes to the site

+++ SpartaIMSWebsite/Controllers/CohortController.cs
@@ -142,9 +142,17 @@ namespace SpartaIMSWebsite.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var cohort = await _context.Cohorts.FindAsync(id);
-            _context.Cohorts.Remove(cohort);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if(cohort.SpartanUsers.Count == 0)
+            {
+                _context.Cohorts.Remove(cohort);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                ViewData["Message"] = "Cannot delete a cohort while users are assigned to it";
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         private bool CohortExists(int id)
